refactor(50): 세션 관련 모듈 경로 통합 및 코드 정리
