+++ src/app/_api/session/session.types.ts
@@ -1,1 +1,1 @@
-export { type SessionData } from "@/lib/session/type";
+export { type SessionData } from "@/lib/session";
