+++ src/app/api/session/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from "next/server";
 
-import { getSessionFromServer } from "@/lib/session/serverSession";
+import { getSessionFromServer } from "@/lib/session";
 
 export const GET = async () => {
   const session = await getSessionFromServer();
