+++ src/lib/api/index.ts
@@ -1,1 +1,2 @@
 export * from "./client";
+export * from "./type";
