+++ src/lib/session/clientSession.ts
@@ -1,6 +1,6 @@
 import { getSession } from "@/app/_api/session/session.api";
 import { TOKEN_TIMES } from "@/constants";
-import { type SessionData } from "@/lib/session/type";
+import { type SessionData } from "@/lib/session";
 
 const clientSessionCache = new Map<
   string,
