+++ src/lib/session/index.ts
@@ -1,3 +1,4 @@
 export { clearClientSessionCache, getSessionFromClient } from "./clientSession";
 export { getSessionFromServer } from "./serverSession";
 export { defaultSession, getSession, sessionOptions } from "./session";
+export { type SessionData } from "./type";
